+++ src/formatters/identity.ts
@@ -1,0 +1,3 @@
+import type { FormatterFunction } from './_types'
+
+export const identity: FormatterFunction = <T>(value: T): T => value