+++ src/formatters/ignore.ts
@@ -1,0 +1,4 @@
+import type { FormatterFunction } from './_types'
+
+// eslint-disable-next-line @typescript-eslint/no-empty-function
+export const ignore: FormatterFunction = (): string => ''