+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # CHANGELOG.md
 
+## 2.22.0 (2021-05-16)
+
+Feature:
+ - add [generateOnlyTypes](https://github.com/ivanhofer/typesafe-i18n#generateOnlyTypes) option
+ - add [identity](https://github.com/ivanhofer/typesafe-i18n#identity) formatter
+ - add [ignore](https://github.com/ivanhofer/typesafe-i18n#ignore) formatter
+
+Readme:
+ - create FAQ section
+
+Fix:
+ - wrap types of Formatters in a `string` to be able to use any kind of formatter name
+
 ## 2.21.1 (2021-05-15)
 
 Fix: